ceph_cfg.mon: Use kwargs to allow to set hostname
Make simple call methods for mon commands

Murge all the model updating code so it only needs to be done once.

Allow hostname applied only if unset

Hostname cant always be gathered correctly for example when the hostname has
changed. Then a hostname for configuring mons needs to be specified.

Signed-off-by: Owen Synge <osynge@suse.com>