ceph_cfg.mon: Reorganize the destroy function
When doing silly things with mon deletion one could previously cause ceph to
upset systemd. We can only remove the mon from the cluster after we have
stopped it. Previously this was true on one node but this needs to be true if
operated in parallel across the cluster. Hence we only remove the mon from the
cluster on the node we remove the mon.

Signed-off-by: Owen Synge <osynge@suse.com>